+++ ChangeLog.md
@@ -11,3 +11,6 @@
 ##### [2023-7-17]
 * new concept RecordTransferredBytes, 对于数据类型T，用户可以通过在tunnel命名空间添加`GetBytes`的T的特化版本，或者添加一个`size_t tunnel_get_bytes() const`成员函数来使统计信息中包含每个节点读取与写出的字节数。
 * 关于中止信息如何跨pipeline传递的方案 [todo]
+
+##### [2023-7-18]
+* add new interface Pipeline.BindExecutorForProcessor